Change the way forces are applied to movers
The way forces are applied to movers is now different.

A force is defined as an object with a type, indicating the
kind of force, and a function which takes a mover in input and
returns a force vector.

The idea is to pass the force functions to the movers, which
will control how the forces are applied depending on their
type (inversion of control).

noc.forces now contains force constructors functions, which take in
account some forces constants and return force objets (also a `custom`
force has been added to simplify new force creation).

`applyForce`, renommed `applyForceVec`, is no longer supposed to be
used directly in the general case. The function `subjectTo`, which
takes a force function is to be used instead. This function allows for a
better control on how forces are applied to a mover (inversion of
control).