IoC on renderer
Renderer is now longer in charge of drawing the movers. It is now passed
to the movers by the simulation, and the movers choose themselves how
they are rendered (IoC).
Consequently, some renderer internal functions, such as shape drawing,
are now part of the renderer object passed to the movers instead.