Add slight repultion between movers
Movers now repulse each over when being too near. The location of the
movers is memorized on the grid, movers look cases around their to know
which other movers are around.
In order to improve performances, add some ugly hack using rendering
steps to reduce the number of iterations. The position of movers on the
field is updated during rendering step.